+++ breealyzer/src/main/java/ch/thenoobs/minecraft/breealyzer/util/allelescoring/scorers/SpeciesAlleleScorer.java
@@ -22,7 +22,7 @@ public class SpeciesAlleleScorer extends AlleleScorer {
 		value = value + getAlleleValue(allele);
 		
 
-		float fValue = value / 2;
+		float fValue = value / (float)2;
 		return fValue;
 	}
